pyproject.toml
[tool.poetry]
name = "api-maribel"
version = "0.1.0"
description = ""
authors = ["JoelFrancisco <joelfranciscofilho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
pytest = "^8.2.2"
requests = "^2.32.3"
django-environ = "^0.11.2"
google-cloud-vision = "^3.7.2"
openai = "^1.34.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"
redis = "^5.0.6"
hiredis = "^2.3.2"
django-cors-headers = "^4.4.0"
gunicorn = "^22.0.0"
whitenoise = "^6.7.0"
boto3 = "^1.34.131"
django-ses = "^4.1.0"
djoser = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
